+++ operations/consumer/DefaultMessageDeserializer.go
@@ -13,6 +13,7 @@ type DefaultMessageDeserializer struct {
 type defaultMessage struct {
 	Partition int32
 	Offset    int64
+	Headers   map[string]string `json:",omitempty" yaml:",omitempty"`
 	Key       *string `json:",omitempty" yaml:",omitempty"`
 	Value     *string
 	Timestamp *time.Time `json:",omitempty" yaml:",omitempty"`
@@ -22,6 +23,7 @@ func (deserializer DefaultMessageDeserializer) newDefaultMessage(msg *sarama.Con
 
 	var key *string
 	var timestamp *time.Time
+	var headers map[string]string
 
 	var value = encodeBytes(msg.Value, flags.EncodeValue)
 
@@ -33,9 +35,14 @@ func (deserializer DefaultMessageDeserializer) newDefaultMessage(msg *sarama.Con
 		timestamp = &msg.Timestamp
 	}
 
+    if flags.PrintHeaders {
+		headers = encodeRecordHeaders(msg.Headers)
+	}
+
 	return defaultMessage{
 		Partition: msg.Partition,
 		Offset:    msg.Offset,
+		Headers:    headers,
 		Key:       key,
 		Value:     value,
 		Timestamp: timestamp,
@@ -48,6 +55,20 @@ func (deserializer DefaultMessageDeserializer) Deserialize(rawMsg *sarama.Consum
 
 	if flags.OutputFormat == "" {
 		var row []string
+
+		if flags.PrintHeaders {
+			if msg.Headers != nil {
+				var column []string
+
+				for key, value := range msg.Headers {
+					column = append(column, key + ":" + value)
+				}
+
+				row = append(row, strings.Join(column[:], ","))
+			} else {
+				row = append(row, "")
+			}
+		}
 
 		if flags.PrintKeys {
 			if msg.Key != nil {