+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add option `print-headers` to print kafka message headers
+- Allow producing multiple messages from stdin
+
 ## 1.2.1 - 2019-05-27
 
 ### Fixed